+++ src/Traits/DefaultValue/UuidBehavior.php
@@ -8,6 +8,11 @@ class UuidBehavior extends DefaultValueBehavior
 {
     public function __construct(string $field)
     {
-        parent::__construct($field, Uuid::uuid4()->toString());
+        parent::__construct($field, null);
+    }
+
+    public function getValue()
+    {
+        return Uuid::uuid4()->toString();
     }
 }